fix: compilation errors + mmcsd stack increase + openclaw build/flash verify
- AP_Filesystem_backend.cpp: fix sign-compare warning (ssize_t vs off_t)
- AP_JSON.cpp: fix sign-compare warning (same pattern)
- lua loadlib.c: mark unused variable i with __attribute__((unused))
- mmcsd_detect thread stack: 2048 -> 4096 (was 90% utilization)
- Verified: build succeeds, flash+verify OK, MAVLink heartbeat + 22 msg types
- RAW_IMU accel confirmed non-zero (xacc=29, yacc=-68, zacc=-991)
- System stable at ~400Hz, CPU load 0.9%

+++ libraries/AP_Filesystem/AP_Filesystem_backend.cpp
@@ -45,7 +45,7 @@ FileData *AP_Filesystem_Backend::load_file(const char *filename)
         delete fd;
         return nullptr;
     }
-    if (read(d, data, st.st_size) != st.st_size) {
+    if ((size_t)read(d, data, st.st_size) != (size_t)st.st_size) {
         close(d);
         free(data);
         delete fd;
